+++ src/lxc/bdev/bdev.h
@@ -128,15 +128,6 @@ bool bdev_destroy(struct lxc_conf *conf);
 /* callback function to be used with userns_exec_1() */
 int bdev_destroy_wrapper(void *data);
 
-/*
- * these are really for qemu-nbd support, as container shutdown
- * must explicitly request device detach.
- */
-bool attach_block_device(struct lxc_conf *conf);
-void detach_block_device(struct lxc_conf *conf);
-
-bool rootfs_is_blockdev(struct lxc_conf *conf);
-
 /* Some helpers for lvm, rdb, and/or loop:
  * Maybe they should move to a separate implementation and header-file
  * (bdev_utils.{c,h}) which can be included in bdev.c?
@@ -147,6 +138,12 @@ int do_mkfs(const char *path, const char *fstype);
 int is_blktype(struct bdev *b);
 int mount_unknown_fs(const char *rootfs, const char *target,
 		const char *options);
-
+bool rootfs_is_blockdev(struct lxc_conf *conf);
+/*
+ * these are really for qemu-nbd support, as container shutdown
+ * must explicitly request device detach.
+ */
+bool attach_block_device(struct lxc_conf *conf);
+void detach_block_device(struct lxc_conf *conf);
 
 #endif // __LXC_BDEV_H
